Create gradle.yml
This uses the "GitHub" wizard for creating a Java/Gradle workflow.

I changed the defaults to run on _all_ `push` and `pull_request` events instead of just those on the `main` branch. I'm guessing that may be important for running checks on development branches, which we definitely want for this project.